requirements.txt
pandas>=1.5.0
numpy>=1.21.0
pyarrow>=10.0.0
akshare>=1.12.0
yfinance>=0.2.18
tushare>=1.2.89
TA-Lib>=0.4.28
scikit-learn>=1.2.0
lightgbm>=4.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
jupyter>=1.0.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
